Initial Commit
P to create a base class Employee with private data members as eid, ename, create a child class 
Programmer that extends base class Employee with data-members as experience, salary and create a child 
class Manager that extends class Employee with datamembers as specializaton, age. Display all the details 
by reading all the data from user. Output should be like this: 
Emp id is: 
Emp Name is: 
Exp is:
Salary is:
Age is: 
Specialization is: